Refactoring + added navigation with Jetpack

+++ app/src/main/java/com/example/composition/domain/entity/GameSettings.kt
@@ -5,8 +5,8 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class GameSettings(
-    val maxSumValue:Int,
-    val minCountOfRightAnswers:Int,
-    val minPercentOfRightAnswers:Int,
-    val gameTimeInSeconds:Int
-):Parcelable
+    val maxSumValue: Int,
+    val minCountOfRightAnswers: Int,
+    val minPercentOfRightAnswers: Double,
+    val gameTimeInSeconds: Int,
+) : Parcelable
